% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_area.R
\name{bg_area}
\alias{bg_area}
\title{BG Area}
\usage{
bg_area(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{strategy}{coordinate strategy}
}
\description{
Calculates the area of a geometry
}
\examples{

}
