% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_centroid.R
\name{bg_centroid}
\alias{bg_centroid}
\title{BG Centroid}
\usage{
bg_centroid(wkt)
}
\arguments{
\item{strategy}{}
}
\description{
Calculates the centroid of geometries
}
\examples{
wkt <- "POLYGON((2 1.3,2.4 1.7,2.8 1.8,3.4 1.2,3.7 1.6,3.4 2,4.1 3,5.3 2.6,5.4 1.2,4.9 0.8,2.9 0.7,2 1.3),(4.0 2.0, 4.2 1.4, 4.8 1.9, 4.4 2.2, 4.0 2.0))"
bg_centroid(wkt)

}
