% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_correct.R
\name{bg_correct}
\alias{bg_correct}
\title{BG Correct}
\usage{
bg_correct(x)
}
\arguments{
\item{x}{well-known text representation of a geometry}
}
\description{
Corrects a geometry: all rings incorrectly oriented with respect to their expected
orientation are reversed. All rings which do not have a closing point
and are typsed as they should have one, the first point is appended.
}
\examples{
bg_correct("POLYGON((0 0, 0 1, 1 1, 1 0))")
bg_correct("POLYGON((0 0, 0 1, 1 1, 1 0),(0.25 0.25, 0.25 0.75, 0.75 0.75, 0.75 0.25))")


}
