% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_crosses.R
\name{bg_crosses}
\alias{bg_crosses}
\title{BG Crosses}
\usage{
bg_crosses(x, y, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{y}{}
}
\description{
Checks if two geometries cross
}
\note{
supported geometries
- linestring
- multi_linestring
}
\examples{
x <- "LINESTRING(-1 -1, 0 0, 1 1)"
y <- c("LINESTRING(1 -1, 0 0, -1 1)", "LINESTRING(0 0, 1 1)", "LINESTRING(-1 1, 0 0, 1 -1)")
bg_crosses(x, y)

}
