% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_densify.R
\name{bg_densify}
\alias{bg_densify}
\title{BG Densify}
\usage{
bg_densify(x, strategy = c("cartesian", "spherical", "geographic"), distance)
}
\arguments{
\item{distance}{numeric. The distance between segments (sequential coordinates)}
}
\description{
Divides segments of a geometry if they are longer than the \code{distance} into
smaller segments
}
\examples{

bg_densify("POLYGON((0 0,0 10,10 10,10 0,0 0),(1 1,4 1,4 4,1 4,1 1))", distance = 6)
bg_densify("POLYGON((0 0,0 10,10 10,10 0,0 0),(1 1,4 1,4 4,1 4,1 1))", strategy = "spherical", distance = 0.06)
bg_densify("POLYGON((0 0,0 10,10 10,10 0,0 0),(1 1,4 1,4 4,1 4,1 1))", strategy = "geographic", distance = 60000)

}
