% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_disjoint.R
\name{bg_disjoint}
\alias{bg_disjoint}
\title{bg_disjoint}
\usage{
bg_disjoint(x, y = NULL, strategy = c("cartesian", "spherical",
  "geographic"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{y}{if null, each element of \code{x} is assessed against every other element of \code{y}}

\item{strategy}{coordinate strategy}
}
\description{
Checks if geometries are spatially disjoint from each other
}
\note{
geographic multipoints are not supported
}
\examples{

bg_disjoint(
  c("LINESTRING(0 0, 1 1)", "LINESTRING(1 1, 2 2)", "POINT(2 2)")
)

bg_disjoint(
  c("LINESTRING(0 0, 1 1)", "LINESTRING(1 1, 2 2)", "POINT(2 2)")
  , c("POINT(1 1)")
)

}
