% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_distance.R
\name{bg_distance}
\alias{bg_distance}
\title{BG Distance}
\usage{
bg_distance(wktFrom, wktTo = NULL, strategy = c("cartesian", "spherical",
  "geographic"))
}
\arguments{
\item{wktTo}{if null, each element of `wktFrom` is assessed against every other element of `wktFrom`}

\item{strategy}{}
}
\value{
matrix
}
\description{
Calculates the distances between two sets of geometries
}
\examples{

}
