% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_envelope.R
\name{bg_envelope}
\alias{bg_envelope}
\title{BG Envelope}
\usage{
bg_envelope(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{x}{}
}
\description{
Calculates the envelope (bounding box) of a geometry
}
