% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_equals.R
\name{bg_equals}
\alias{bg_equals}
\title{BG Equals}
\usage{
bg_equals(x, y)
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{y}{well-known text representation of a geometry}
}
\value{
Logical matrix returning \code{TRUE} if \code{x} equals \code{y}
}
\description{
Checks if geometries are spatially equal
}
\examples{

bg_equals(x = "POINT(0 0)", y = "POINT(0 0)")
bg_equals(x = "LINESTRING(0 0,1 1,2 2)", y = "LINESTRING(0 0,1 1,2 2)")
bg_equals(
  x = c("POLYGON((0 0,0 1,1 1,1 0,0 0))", "MULTILINESTRING((0 0,2 2))")
  , y = c("POINT(1 1)", "LINESTRING(0 0,2 2)", "MULTIPOLYGON(((0 0,0 1,1 1,1 0,0 0)))")
  )

}
