% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_intersects.R
\name{bg_intersects}
\alias{bg_intersects}
\title{bg_intersects}
\usage{
bg_intersects(x, y = NULL, strategy = c("cartesian", "spherical"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{y}{if null, each element of \code{x} is assessed against every other element of \code{y}}

\item{strategy}{coordinate strategy}
}
\description{
Checks if geometries are spatially disjoint from each other
}
\note{
cartesian and spherical strategies supported
}
\examples{

bg_intersects(
  c("LINESTRING(0 0, 1 1)", "LINESTRING(1 1, 2 2)", "POINT(2 2)")
)

bg_intersects(
  c("LINESTRING(0 0, 1 1)", "LINESTRING(1 1, 2 2)", "POINT(2 2)")
  , c("POINT(1 1)")
)

}
