% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_is.R
\name{bg_is_empty}
\alias{bg_is_empty}
\title{BG Is Empty}
\usage{
bg_is_empty(wkt, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{strategy}{}
}
\description{
Checks if a geometry is empty
}
\examples{

}
