% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_is.R
\name{bg_is_simple}
\alias{bg_is_simple}
\title{BG Is Simple}
\usage{
bg_is_simple(wkt, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{wktFrom}{}
}
\description{
Checks if a geometry is simple
}
\examples{

}
