% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_is.R
\name{bg_is_valid}
\alias{bg_is_valid}
\title{BG Is Valid}
\usage{
bg_is_valid(wkt, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{strategy}{}
}
\description{
Checks if a geometry is valid
}
\examples{

}
