% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_comparable_distance.R
\name{bg_nearest}
\alias{bg_nearest}
\title{BG Nearest}
\usage{
bg_nearest(x, y, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{y}{if null, each element of \code{x} is assessed against every other element of \code{y}}

\item{strategy}{coordinate strategy}
}
\value{
Numeric vector the same length of \code{x}, each element being the index
of \code{y} containing the nearest geometry
}
\description{
Calculates the comparable distance measurement of multiple geometries and returns
the nearest
}
\examples{
bg_nearest(
  x = c("POINT(0 0)", "POINT(-2 -2)", "POINT(2 2)", "POINT(1 1)", "POINT(-1 1)", "POINT(-4 0)"),
  y = c("POINT(0 0)", "POINT(-2 -2)","POINT(1 1)")
  )

bg_nearest(
  x = c("LINESTRING(-1 -1, 1 1)","POLYGON((-1 -1, 1 -1, 1 1, -1 1, -1 -1))"),
  y = c("LINESTRING(-1 -2, 1 -2)","LINESTRING(0 2, 2 0.5)")
 )


}
