% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_num.R
\name{bg_number_geometries}
\alias{bg_number_geometries}
\title{BG Number Geometries}
\usage{
bg_number_geometries(wkt)
}
\arguments{
\item{wkt}{}
}
\description{
Calculates the number of geometries of a geometry
}
\examples{

}
