% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_num.R
\name{bg_number_interior_rings}
\alias{bg_number_interior_rings}
\title{BG Number Interior Rings}
\usage{
bg_number_interior_rings(wkt)
}
\arguments{
\item{wkt}{}
}
\description{
Calculates the number of interior rings of a geometry
}
\examples{

}
