% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_num.R
\name{bg_number_points}
\alias{bg_number_points}
\title{BG Number Points}
\usage{
bg_number_points(wkt)
}
\arguments{
\item{wkt}{}
}
\description{
Calculates the number of points of a geometry
}
\examples{

}
