% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_num.R
\name{bg_number_segments}
\alias{bg_number_segments}
\title{BG Number Segments}
\usage{
bg_number_segments(wkt)
}
\arguments{
\item{wkt}{}
}
\description{
Calculates the number of segments of a geometry
}
\examples{

}
