% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_perimeter.R
\name{bg_perimeter}
\alias{bg_perimeter}
\title{BG Perimeter}
\usage{
bg_perimeter(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{strategy}{}
}
\description{
Calculates the length of the perimeter of a geometry
}
