% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_reverse.R
\name{bg_reverse}
\alias{bg_reverse}
\title{BG Reverse}
\usage{
bg_reverse(x)
}
\arguments{
\item{x}{}
}
\description{
Reverses the points in a geometry
}
\examples{
bg_reverse("POLYGON((0 0,0 1,1 1,1 0,0 0))")

}
