% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_simplify.R
\name{bg_simplify}
\alias{bg_simplify}
\title{BG Simplify}
\usage{
bg_simplify(x, distance = 0.5, strategy = c("cartesian", "spherical"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{distance}{distance between segments}

\item{strategy}{coordinate strategy}
}
\description{
Simplifies a geometry
}
\note{
Uses the Douglas-Peucker algorithm

Supported geometries
- polygon
}
\examples{

polygon <- "POLYGON((0 0, 0 0.2, 0.2 0.2, 0.2 0, 0.3 0, 0.3 0.2, 0.25, 0.5, 0.1 0.4, 0 0.2, 0 0))"
bg_simplify(polygon, distance = 0.2)


}
