% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_unique.R
\name{bg_unique}
\alias{bg_unique}
\title{BG Unique}
\usage{
bg_unique(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{strategy}{}
}
\description{
Calculates the minimal set (where duplicate consecutive points are removed) of a geometry.
}
