% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeometry-package.R
\docType{package}
\name{boostgeometry-package}
\alias{boostgeometry}
\alias{boostgeometry-package}
\title{Provides access to the 'Boost Geometry' 'C++' header-only library.}
\description{
Provides access to the 'Boost Geometry' 'C++' header-only library.
}
\examples{
\dontrun{
# Include a dependency to boostgeometry inside a cpp file
// [[Rcpp::depends(boostgeometry)]]
include <Rcpp.h>
}
}
