% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_envelope.R
\name{bg_envelope}
\alias{bg_envelope}
\title{BG Envelope}
\usage{
bg_envelope(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{strategy}{coordinate strategy}
}
\description{
Calculates the envelope (bounding box) of a geometry
}
\examples{
polygon <- "POLYGON((2 1.3,2.4 1.7,2.8 1.8,3.4 1.2,3.7 1.6,3.4 2,4.1 3,5.3 2.6,5.4 1.2,4.9 0.8,2.9 0.7,2 1.3), (4.0 2.0, 4.2 1.4, 4.8 1.9, 4.4 2.2, 4.0 2.0))"
bg_envelope(polygon)

bg_envelope("LINESTRING(144 -37, 145 -36)", "geographic")

}
