% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_is.R
\name{bg_is_empty}
\alias{bg_is_empty}
\title{BG Is Empty}
\usage{
bg_is_empty(x)
}
\arguments{
\item{x}{well-known text representation of a geometry}
}
\description{
Checks if a geometry is empty
}
\examples{
bg_is_empty("MULTILINESTRING((0 0,0 10,10 0),(1 1,8 1,1 8))")

}
