% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_is.R
\name{bg_is_simple}
\alias{bg_is_simple}
\title{BG Is Simple}
\usage{
bg_is_simple(x)
}
\arguments{
\item{x}{well-known text representation of a geometry}
}
\description{
Checks if a geometry is simple
}
\examples{
bg_is_simple("MULTILINESTRING((0 0,0 10,10 10,10 0,0 0),(10 10,20 20))")
bg_is_simple("LINESTRING(0 0,0 10,10 10,10 0,0 0)")

}
