% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{bg_join}
\alias{bg_join}
\title{bg join}
\usage{
bg_join(x, y, join = "intersects")
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{y}{if null, each element of \code{x} is assessed against every other element of \code{y}}
}
\description{
Joins geometries
}
