% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_length.R
\name{bg_length}
\alias{bg_length}
\title{BG Length}
\usage{
bg_length(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{strategy}{coordinate strategy}
}
\description{
Calculates the length of a geometry
}
\note{
points - returns 0
lines - returns the length
areal - returns 0
}
\examples{
bg_length("LINESTRING(0 0,1 1,4 8,3 2)")
bg_length("LINESTRING(0 0,1 1,4 8,3 2)", "geographic")

}
