% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_length.R
\name{bg_length}
\alias{bg_length}
\title{BG Length}
\usage{
bg_length(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{strategy}{}
}
\description{
Calculates the length of a geometry
}
\examples{
bg_length("LINESTRING(0 0,1 1,4 8,3 2)")
bg_length("LINESTRING(0 0,1 1,4 8,3 2)", "geographic")

}
