% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_num.R
\name{bg_number_interior_rings}
\alias{bg_number_interior_rings}
\title{BG Number Interior Rings}
\usage{
bg_number_interior_rings(x)
}
\arguments{
\item{x}{}
}
\description{
Calculates the number of interior rings of a geometry
}
\examples{
bg_number_interior_rings("POLYGON((0 0,0 10,10 0,0 0),(1 1,1 2,2 1,1 1),(3 3,3 4,4 3,3 3))")

}
