% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_num.R
\name{bg_number_segments}
\alias{bg_number_segments}
\title{BG Number Segments}
\usage{
bg_number_segments(x)
}
\arguments{
\item{x}{well-known text representation of a geometry}
}
\description{
Calculates the number of segments of a geometry
}
\examples{
bg_number_segments("MULTIPOLYGON(((0 0,0 10,10 0),(1 1,8 1,1 8)),((10 10,10 20,20 10)))")

}
