% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_unique.R
\name{bg_unique}
\alias{bg_unique}
\title{BG Unique}
\usage{
bg_unique(x, strategy = c("cartesian", "spherical", "geographic"))
}
\arguments{
\item{x}{well-known text representation of a geometry}

\item{strategy}{coordinate strategy}
}
\description{
Calculates the minimal set (where duplicate consecutive points are removed) of a geometry.
}
\note{
Supported geometries
- linestring
- multilinestring
- polygon
- multipolygon
}
\examples{
bg_unique("LINESTRING(0 0,0 0,1 1,1 1,2 2)")
bg_unique("POLYGON((0 0,0 0,0 5,5 5,5 5,5 5,5 0,5 0,0 0,0 0,0 0,0 0))")

}
